+++ sizemeup/cli/build.py
@@ -3,6 +3,7 @@ import logging
 import os
 import sys
 from pathlib import Path
+from statistics import mean
 
 import rich
 import rich.console
@@ -13,7 +14,11 @@ from rich.logging import RichHandler
 from rich.table import Table
 
 import sizemeup
-from sizemeup.ncbi import get_genome_sizes
+from sizemeup.atb import download_atb_file, parse_assembly_stats
+from sizemeup.ncbi import get_genome_sizes, species2taxid
+
+from bactopia.atb import parse_atb_file_list
+from bactopia.utils import download_url, file_exists, validate_file
 
 
 # Set up Rich
@@ -33,6 +38,14 @@ click.rich_click.OPTION_GROUPS = {
             "options": [
                 "--ncbi-api-key",
                 "--chunk-size",
+            ],
+        },
+        {
+            "name": "ATB Options",
+            "options": [
+                "--atb-file-list-url",
+                "--atb-assembly-stats",
+                "--min-genomes",
             ],
         },
         {
@@ -68,8 +81,28 @@ click.rich_click.OPTION_GROUPS = {
 @click.option(
     "--chunk-size",
     "-c",
-    default=200,
+    default=500,
     help="The size of the chunks to split the list into",
+)
+@click.option(
+    "--atb-file-list-url",
+    "-a",
+    default="https://osf.io/download/4yv85/",
+    show_default=True,
+    help="The URL to the ATB file list",
+)
+@click.option(
+    "--atb-assembly-stats-url",
+    "-s",
+    default="https://osf.io/download/nbyqv/",
+    show_default=True,
+    help="The name of the ATB assembly stats file",
+)
+@click.option(
+    "--min-genomes",
+    "-m",
+    default=5,
+    help="The minimum number of genomes to use for ATB species",
 )
 @click.option(
     "--user-sizes",
@@ -84,6 +117,9 @@ def sizemeup_build(
     outdir: str,
     ncbi_api_key: str,
     chunk_size: int,
+    atb_file_list_url: str,
+    atb_assembly_stats_url: str,
+    min_genomes: int,
     user_sizes: str,
     force: bool,
     verbose: bool,
@@ -107,6 +143,7 @@ def sizemeup_build(
     Path(outdir).mkdir(parents=True, exist_ok=True)
 
     user_genome_sizes = {}
+    species_list = {}
     if user_sizes:
         logging.info(f"Reading user-provided genome sizes from {user_sizes}")
         with open(user_sizes, "r") as f:
@@ -129,6 +166,53 @@ def sizemeup_build(
     for taxid, genome in user_genome_sizes.items():
         if taxid not in genome_sizes:
             genome_sizes[taxid] = genome
+
+    # Create species list to filter out of ATB
+    for taxid, genome in genome_sizes.items():
+        species_list[genome["name"]] = taxid
+    logging.info(f"Found {len(species_list)} species from NCBI/User to filter from ATB")
+
+    # Add ATB sizes to the genome sizes, only if they don't already exist
+    # NCBI sizes and user sizes will ALWAYS take precedence
+    logging.debug("Getting assembly sizes from ATB")
+    atb_file_list = download_atb_file(
+        f"{str(outdir)}/file_list.all.latest.tsv.gz",
+        atb_file_list_url,
+        progress=False if silent else True,
+    )
+    atb_assembly_stats = download_atb_file(
+        f"{str(outdir)}/assembly-stats.tsv.gz",
+        atb_assembly_stats_url,
+        progress=False if silent else True,
+    )
+
+    # Parse the ATB file list
+    logging.info(f"Parsing ATB file list: {atb_file_list}")
+    samples, archives, species = parse_atb_file_list(atb_file_list)
+    assembly_stats = parse_assembly_stats(atb_assembly_stats)
+
+    # Capture ATB sizes for species that don't already have a size
+    atb_species = {}
+    for name, samples in species.items():
+        if name not in species_list:
+            if "_" not in name or not any(i.isdigit() for i in name):
+                if len(samples) >= min_genomes:
+                    atb_species[name] = samples
+            else:
+                logging.debug(f"Skipping {name} due to underscore or digit in name")
+
+    species_taxids = species2taxid(list(atb_species.keys()), ncbi_api_key, chunk_size)
+    for name, samples in atb_species.items():
+        if name in species_taxids:
+            taxid = species_taxids[name]
+            genome_sizes[taxid] = {
+                "name": name,
+                "taxid": taxid,
+                "category": "bacteria",
+                "expected_ungapped_length": int(mean([int(assembly_stats[sample]["total_length"]) for sample in samples])),
+                "source": "atb",
+                "method_determined": f"Average assembly size of {len(samples)} AllTheBacteria samples",
+            }
 
     # Write the genome sizes to a file
     logging.info(f"Writing genome sizes to {outdir}/sizemeup-sizes.txt")